INC-1143: Refactor large resource class into two (#390)
Split incentive levels from prison incentive levels into separate resources
